+++ pace-util/HISTORY.md
@@ -10,6 +10,7 @@ Major changes:
 Minor changes:
 - updated QuantityFactory to accept the more generic GridSizer class on initialization
 - added `sizer` as public attribute on QuantityFactory
+- added `Namelist` class to initialize namelist files used in fv3gfs-fortran
 
 v0.7.0
 ------